+++ pom.xml
@@ -164,7 +164,7 @@
 		<swt.version>3.133.0</swt.version>
 		<swt.svg.version>3.132.0</swt.svg.version>
 		<jsvg.version>2.1.0</jsvg.version>
-		<lombok.version>1.18.44</lombok.version>
+		<lombok.version>1.18.46</lombok.version>
 		<xmlunit.version>2.11.0</xmlunit.version>
 		<org.eclipse.ui.workbench.version>3.138.0</org.eclipse.ui.workbench.version>
 		<org.eclipse.osgi.version>3.24.100</org.eclipse.osgi.version>
