+++ src/main/java/cl/cavallinux/jisocreator/gui/dialog/ShowIsoLayoutInformationDialog.java
@@ -17,7 +17,7 @@ import org.eclipse.swt.widgets.Text;
 import cl.cavallinux.jisocreator.gui.i18n.ShowIsoInformationDialogMessages;
 import cl.cavallinux.jisocreator.gui.listeners.dialog.EnterKeySubmitAdapter;
 import cl.cavallinux.jisocreator.model.isoexplorer.impl.IsoFileSystem;
-import cl.cavallinux.jisocreator.model.parser.IsoFilesystemParser;
+import cl.cavallinux.jisocreator.model.parser.decl.IsoFilesystemParser;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.Setter;
